Move class to cmd_entries/args.py
