+++ sourced/ml/cmd_entries/__init__.py
@@ -1,3 +1,4 @@
+from sourced.ml.cmd_entries.args import ArgumentDefaultsHelpFormatterNoNone
 from sourced.ml.cmd_entries.bigartm2asdf import bigartm2asdf_entry
 from sourced.ml.cmd_entries.bow_converters import bow2vw_entry
 from sourced.ml.cmd_entries.dump_model import dump_model
